+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+### Version 4.22.0 (5th June 2020)
+#### Added
+- Added subscription tracking feature.
+
+#### Native SDKs
+- [iOS@v4.22.0][ios_sdk_v4.22.0]
+- [Android@v4.22.0][android_sdk_v4.22.0]
+- [Windows@v4.17.0][windows_sdk_v4.17.0]
+
+---
+
 ### Version 4.21.3 (4th May 2020)
 #### Fixed
 - Removed iAd timer from iOS native SDK.
@@ -721,6 +732,7 @@
 [ios_sdk_v4.21.0]: https://github.com/adjust/ios_sdk/tree/v4.21.0
 [ios_sdk_v4.21.1]: https://github.com/adjust/ios_sdk/tree/v4.21.1
 [ios_sdk_v4.21.2]: https://github.com/adjust/ios_sdk/tree/v4.21.2
+[ios_sdk_v4.22.0]: https://github.com/adjust/ios_sdk/tree/v4.22.0
 
 [android_sdk_v3.5.0]: https://github.com/adjust/android_sdk/tree/v3.5.0
 [android_sdk_v4.1.0]: https://github.com/adjust/android_sdk/tree/v4.1.0
@@ -747,6 +759,7 @@
 [android_sdk_v4.20.0]: https://github.com/adjust/android_sdk/tree/v4.20.0
 [android_sdk_v4.21.0]: https://github.com/adjust/android_sdk/tree/v4.21.0
 [android_sdk_v4.21.1]: https://github.com/adjust/android_sdk/tree/v4.21.1
+[android_sdk_v4.22.0]: https://github.com/adjust/android_sdk/tree/v4.22.0
 
 [windows_sdk_v4.12.0]: https://github.com/adjust/windows_sdk/tree/v4.12.0
 [windows_sdk_v4.13.0]: https://github.com/adjust/windows_sdk/tree/v4.13.0